% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{coef.demogfit_fit}
\alias{coef.demogfit_fit}
\title{Extract fitted parameters from a demogfit_fit object}
\usage{
\method{coef}{demogfit_fit}(object, ...)
}
\arguments{
\item{object}{a \code{demogfit_fit} object, as returned by
\code{\link{fit_si}}, \code{\link{fit_im}}, \code{\link{fit_iim}} or
\code{\link{fit_sc}}}

\item{...}{unused, present for S3 consistency}
}
\value{
a named numeric vector of fitted parameter estimates
}
\description{
Extract fitted parameters from a demogfit_fit object
}
