% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{.nll}
\alias{.nll}
\title{Composite negative log-likelihood, generic across models}
\usage{
.nll(par, s_dist, model_fn, fixed_args = list())
}
\description{
Composite negative log-likelihood, generic across models
}
\keyword{internal}
