% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{.partial_exp_sum}
\alias{.partial_exp_sum}
\title{Partial Poisson sum: sum_{l=0}^{k} lambda^l / l!}
\usage{
.partial_exp_sum(lambda, k)
}
\description{
Partial Poisson sum: sum_{l=0}^{k} lambda^l / l!
}
\keyword{internal}
