% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{.upper_incomplete_gamma}
\alias{.upper_incomplete_gamma}
\title{Upper incomplete gamma function Gamma(s, x), via R's regularized lower
incomplete gamma (pgamma).}
\usage{
.upper_incomplete_gamma(s, x)
}
\description{
Upper incomplete gamma function Gamma(s, x), via R's regularized lower
incomplete gamma (pgamma).
}
\keyword{internal}
