% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{.wh_lambdas}
\alias{.wh_lambdas}
\title{Eigenvalues of the 2-deme migration matrix used throughout WH's solutions.}
\usage{
.wh_lambdas(M)
}
\description{
Eigenvalues of the 2-deme migration matrix used throughout WH's solutions.
}
\keyword{internal}
